Deadshot Late Night Coding
Listen, it was not a good idea for me to code this late at night, but I did it anyways.
I'm going to get it functional, and clean-up is a future Jordan problem that will be solved. Right now, this is an organisational disaster.
I still have no idea how Cameron's aiming animation stuff works, so that's a bit borked.