Faith System Added
Added faith regeneration rate and it should restore on the sword hitting the boss.
Currently it doesn't regenerate while swinging the sword, but that can be added with one line of code.
Also moved the HUD manager out of the Player States folder because I don't know why it was there.